Reworked auto direction algorithm, prevention from popover going off screen
