+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "react-native-popover",
-  "version": "0.4.1",
+  "version": "0.4.2",
   "description": "A <Popover /> component for react-native",
-  "main": "NewPopover.js",
+  "main": "Popover.js",
   "author": "Jean Regisser <jean.regisser@gmail.com> (https://github.com/jeanregisser)",
   "keywords": [
     "react-component",
